Update patch set 2
Patch Set 2: Fails

Indeed, a similar change exists in our master branch already, such that we won't be able to merge this change all the way. There's also a conflict with the current open-source master, so even if we want to merge it some work needs to be done to resolve the conflict.

Christian, it's up to you: you can abandon that change, or you can upload a replacement to resolve the conflict so that we can submit it, but you need to understand that we'll have to ignore your change into our internal tree, i.e. that your change will only live in the open-source master tree until it gets overwritten by the code we've written instead.

Patch-set: 2
Reviewer: Jean-Baptiste Queru <1000660@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=-1